Make it Pretty, Found Bug
So I made it a little more user friendly and while I did so I found that the properties on my hero object are being shared with my hero default. This makes sense because prototype. I must figure out how to create a seperate instance of the Hero class for hero and defaultHero.